allow yaml parser and visitor to be garbage collected
When loading a YAML dictionary the YamlFile class will store a YAML visitor and a parser.

This means that after "read_file_into_dictionary" is called the Psych:Parser will keep the object tree that was used to parse and navigate the YAML document, even after the resulting Hash dictionary is created.

While the memory required to parse the YAML dictionary is still required, we can change the YamlFile class to allow the Psych::Parser garbage collection, freeing a significant amount of memory from the heap.