implementation of streaming yaml parsing (#106)
Instead of running Psych::Parse which creates the entire Psych-based Tree before emitting the ruby hash dictionary, the plugin should be capable of gradually constructing the final hash dictionary as the parser identifies YAML elements (streaming parsing).

This changeset adds an opt-in streaming parsing for YAML files based on snakeyaml-engine. When enabled, objects created during the parsing of the YAML will be immediately be available for collection, as opposed to being reachable until the parsing ends. The most significant impact will be observed with multiple translate filters performing load/refresh at the same time, where the GC is able to collect objects as they're processing the YAML files, avoiding potential OOMs.

Co-authored-by: Andrea Selva <selva.andre@gmail.com>